- change UID of variable to int - name thread (TimeCachedThread + Time limit) - add concurrent propagation but there are still some bugs...
git-svn-id: https://campus.mines-nantes.fr/contraintes/galak/trunk@382 031d3fc5-a07f-0410-aa2d-fafe74bc0a1b

+++ solver/src/main/java/solver/variables/BoolVarImpl.java
@@ -75,7 +75,7 @@ public final class BoolVarImpl implements BoolVar {
 
     protected HeuristicVal heuristicVal;
 
-    protected long uniqueID;
+    protected int uniqueID;
 
     protected final IPropagationEngine engine;
 
@@ -88,11 +88,11 @@ public final class BoolVarImpl implements BoolVar {
         requests = RequestListBuilder.preset(solver.getEnvironment());
     }
 
-    public long getUniqueID() {
+    public int getUniqueID() {
         return uniqueID;
     }
 
-    public void setUniqueID(long uniqueID) {
+    public void setUniqueID(int uniqueID) {
         this.uniqueID = uniqueID;
     }
